--- src/main/java/com/idata/eboks/configs/templates/EboksRestTemplate.java
@@ -1,29 +1,0 @@
-package com.idata.eboks.configs.templates;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class EboksRestTemplate {
-
-    @Value("$eboks.api.key")
-    private String apiKey;
-
-    @Bean
-    @Qualifier("eBokRestTemplate")
-    public RestTemplate restTemplate() {
-
-        RestTemplate restTemplate = new RestTemplate();
-
-        restTemplate.getInterceptors().add((request, body, execution) -> {
-            request.getHeaders().add("Authorization", "Bearer " + apiKey);
-            return execution.execute(request, body);
-        });
-        return restTemplate;
-
-    }
-
-}
