+++ src/main/java/com/idata/eboks/controller/UserController.java
@@ -1,13 +1,17 @@
 package com.idata.eboks.controller;
 
+import com.idata.eboks.models.EndUser;
 import com.idata.eboks.models.UserMatch;
 import com.idata.eboks.Services.UserService;
+
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/v1/tenant/{tenantKey}/users")
+@RequestMapping("/v1/tenant/")
 public class UserController {
 
     private final UserService userService;
@@ -17,8 +21,13 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping
+    @PostMapping("/{tenantKey}/usermatch")
     public ResponseEntity<UserMatch> userMatch(@PathVariable String tenantKey, @RequestBody UserMatch userMatch) {
         return ResponseEntity.ok(userService.matchUsers(tenantKey,userMatch));
+    }
+        // Kollar alla mottagare för tenants
+    @GetMapping("/{tenantKey}/users")
+    public ResponseEntity<List<EndUser>> findAllUsers(@PathVariable String tenantKey) {
+        return ResponseEntity.ok(userService.findUsersFromTenants(tenantKey));
     }
 }
