+++ src/main/java/com/idata/eboks/controller/UserMatchController.java
@@ -1,84 +1,84 @@
-package com.idata.eboks.controller;
+// package com.idata.eboks.controller;
 
-import java.util.List;
-import java.util.Map;
+// import java.util.List;
+// import java.util.Map;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.beans.factory.annotation.Value;
+// import org.springframework.http.ResponseEntity;
+// import org.springframework.web.bind.annotation.*;
 
-import com.idata.eboks.Services.UserMatchService;
-import com.idata.eboks.models.ContentUser;
-import com.idata.eboks.models.EndUser;
-import com.idata.eboks.models.Tenant;
-import com.idata.eboks.models.TenantKey;
-import com.idata.eboks.models.UserMatch;
+// import com.idata.eboks.Services.UserMatchService;
+// import com.idata.eboks.models.ContentUser;
+// import com.idata.eboks.models.EndUser;
+// import com.idata.eboks.models.Tenant;
+// import com.idata.eboks.models.TenantKey;
+// import com.idata.eboks.models.UserMatch;
 
-@RestController
-@RequestMapping("/v1/tenant")
-public class UserMatchController {
+// @RestController
+// @RequestMapping("/v1/tenant")
+// public class UserMatchController {
 
-    @Value("${billo.api.client.id}")
-    private String clientId;
+//     @Value("${billo.api.client.id}")
+//     private String clientId;
 
-    @Value("${billo.api.client.secret}")
-    private String clientSecret;
+//     @Value("${billo.api.client.secret}")
+//     private String clientSecret;
 
-    @Value("${billo.api.client.tenantKey}")
-    private String tenantKey;
+//     @Value("${billo.api.client.tenantKey}")
+//     private String tenantKey;
 
-    @Autowired
-    private UserMatchService userMatchService;
+//     @Autowired
+//     private UserMatchService userMatchService;
 
-    // Kollar alla mottagare för tenants
-    @GetMapping("/{tenantKey}/users")
-    public ResponseEntity<List<EndUser>> findAllUsers(@PathVariable String tenantKey) {
-        return ResponseEntity.ok(userMatchService.findUsersFromTenants(tenantKey));
-    }
+//     // Kollar alla mottagare för tenants
+//     @GetMapping("/{tenantKey}/users")
+//     public ResponseEntity<List<EndUser>> findAllUsers(@PathVariable String tenantKey) {
+//         return ResponseEntity.ok(userMatchService.findUsersFromTenants(tenantKey));
+//     }
 
-    // Kollar om mottagare kan få meddelande från tenant
-    @PostMapping("/{tenantKey}/usermatch")
-    public ResponseEntity<UserMatch> userMatch(@PathVariable String tenantKey, @RequestBody UserMatch userMatch) {
+//     // Kollar om mottagare kan få meddelande från tenant
+//     @PostMapping("/{tenantKey}/usermatch")
+//     public ResponseEntity<UserMatch> userMatch(@PathVariable String tenantKey, @RequestBody UserMatch userMatch) {
 
-        return ResponseEntity.ok(userMatchService.matchUsers(tenantKey, userMatch));
-    }
+//         return ResponseEntity.ok(userMatchService.matchUsers(tenantKey, userMatch));
+//     }
 
-    @PostMapping("/{tenantKey}/content")
-    public ResponseEntity<ContentUser> sendcontent(@PathVariable String tenantKey, @RequestBody ContentUser contentUser) {
+//     @PostMapping("/{tenantKey}/content")
+//     public ResponseEntity<ContentUser> sendcontent(@PathVariable String tenantKey, @RequestBody ContentUser contentUser) {
 
-        System.out.println("--------------  här kommer Meddelandet ------------");
-        System.out.println(contentUser);
-        System.out.println("--------------  ------------------ ------------");
+//         System.out.println("--------------  här kommer Meddelandet ------------");
+//         System.out.println(contentUser);
+//         System.out.println("--------------  ------------------ ------------");
 
-        return ResponseEntity.ok(userMatchService.sendContentToUser(tenantKey, contentUser));
-    }
+//         return ResponseEntity.ok(userMatchService.sendContentToUser(tenantKey, contentUser));
+//     }
 
-    // Skapar en ny tenant
-    @PostMapping
-    public ResponseEntity<Tenant> createTenant(@RequestBody Tenant tenant) {
-        System.out.println("Received tenant: " + tenant);
-        Tenant createdTenant = userMatchService.createTenant(tenant);
-        return ResponseEntity.status(createdTenant != null ? 201 : 200).body(createdTenant);
-    }
+//     // Skapar en ny tenant
+//     @PostMapping
+//     public ResponseEntity<Tenant> createTenant(@RequestBody Tenant tenant) {
+//         System.out.println("Received tenant: " + tenant);
+//         Tenant createdTenant = userMatchService.createTenant(tenant);
+//         return ResponseEntity.status(createdTenant != null ? 201 : 200).body(createdTenant);
+//     }
 
-    // Listar alla tenants
-    @GetMapping
-    public ResponseEntity<List<List<TenantKey>>> listTenants(@RequestParam(required = false) String orgnr) {
-        System.out.println("listed tenant: " + orgnr);
-        List<List<TenantKey>> tenants = userMatchService.listTenants(orgnr);
-        return ResponseEntity.ok(tenants);
-    }
+//     // Listar alla tenants
+//     @GetMapping
+//     public ResponseEntity<List<List<TenantKey>>> listTenants(@RequestParam(required = false) String orgnr) {
+//         System.out.println("listed tenant: " + orgnr);
+//         List<List<TenantKey>> tenants = userMatchService.listTenants(orgnr);
+//         return ResponseEntity.ok(tenants);
+//     }
 
-    @GetMapping("{tenantKey}/test")
-    public ResponseEntity<ContentUser> test(@PathVariable String tenantKey, @RequestBody ContentUser contentUser) {
-        return ResponseEntity.ok(contentUser);
-    }
+//     @GetMapping("{tenantKey}/test")
+//     public ResponseEntity<ContentUser> test(@PathVariable String tenantKey, @RequestBody ContentUser contentUser) {
+//         return ResponseEntity.ok(contentUser);
+//     }
 
-    // Ändrar namn på tenant
-    @GetMapping("{tenantKey}/name")
-    public ResponseEntity<Tenant> updateTenantName(@PathVariable String tenantKey) {
-        return ResponseEntity.ok(userMatchService.updateTenantName(tenantKey, "Bosse"));
-    }
+//     // Ändrar namn på tenant
+//     @GetMapping("{tenantKey}/name")
+//     public ResponseEntity<Tenant> updateTenantName(@PathVariable String tenantKey) {
+//         return ResponseEntity.ok(userMatchService.updateTenantName(tenantKey, "Bosse"));
+//     }
 
-}
+// }
